move analytics data fetch up to campaign component
this is so the data can be used to calculate CTRs in the CTA component which is a sibling of the analytics component
